Version 0.1
Functionality:
Asks on startup for username for the wallpaper fetching progress.
Fetch cookies if logged in browser to steam so it can fetch wallpapers that user is subscribed to.
Caches ids and thumbnail pictures of the wallpapers so the startup is faster after first.
Change wallpaper per display.
Auto queue wallpapers for a display.

Known issues:
Setting same wallpaper for multiple monitors is not working properly.
Queue function doesn't work properly if multiple displays are selected.